# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddChatMemberDTO:
    properties:
      role:
        description: Роль пользователя в чате - админ(писать и добавлять участников),
          участник(писать), зритель (только просмотр)
        type: string
      user_id:
        format: uuid
        type: string
    type: object
  dto.AddUsersToChatDTO:
    properties:
      members:
        items:
          $ref: '#/definitions/dto.AddChatMemberDTO'
        type: array
    type: object
...
  dto.ChatCreateInformationDTO:
    properties:
      members:
        items:
          $ref: '#/definitions/dto.AddChatMemberDTO'
        type: array
      name:
        type: string
      type:
        description: Тип чата - канал, диалог или группа
        type: string
    type: object
  dto.ChatDetailedInformationDTO:
    properties:
      can_chat:
        type: boolean
      description:
        type: string
      id:
        format: uuid
        type: string
      is_admin:
        type: boolean
      is_member:
        type: boolean
      is_private:
        type: boolean
      members:
        items:
          $ref: '#/definitions/dto.UserInfoChatDTO'
        type: array
      messages:
        items:
          $ref: '#/definitions/dto.MessageDTO'
        type: array
      name:
        type: string
      type:
        type: string
    type: object
  dto.ChatUpdateDTO:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  dto.ChatViewInformationDTO:
    properties:
      id:
        format: uuid
        type: string
      last_message:
        type: object
      name:
        type: string
      type:
        type: string
    type: object
...
  dto.ErrorDTO:
    properties:
      message:
        type: string
    type: object
  dto.GetAvatarsRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
  dto.GetAvatarsResponse:
    properties:
      avatars:
        additionalProperties:
          type: string
        description: map[id]url
        type: object
    type: object
...
  dto.IdDTO:
    properties:
      id:
        format: uuid
        type: string
    type: object
...
  dto.MessageDTO:
    properties:
      chat_id:
        format: uuid
        type: string
      created_at:
        format: date-time
        type: string
      id:
        format: uuid
        type: string
      sender_id:
        format: uuid
        type: string
      sender_name:
        type: string
      text:
        type: string
      type:
        description: Тип сообщения - системное или пользовательское
        type: string
      updated_at:
        format: date-time
        type: string
    type: object
...
  dto.UserInfoChatDTO:
    properties:
      role:
        description: Роль пользователя в чате - админ(писать и добавлять участников),
          участник(писать), зритель (только просмотр)
        type: string
      user_id:
        format: uuid
        type: string
      user_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Undefined Team
    url: https://github.com/go-park-mail-ru/2025_2_Undefined
  description: API сервер для чат-приложения в стиле Telegram. Позволяет регистрировать
    пользователей, управлять чатами и обмениваться сообщениями.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Undefined team API documentation of project Telegram
  version: "1.0"
paths:
  /chats:
    get:
      consumes:
      - application/json
      description: Возвращает список всех чатов текущего пользователя с информацией
        о последнем сообщении
      produces:
      - application/json
      responses:
        "200":
          description: Список чатов
          schema:
            items:
              $ref: '#/definitions/dto.ChatViewInformationDTO'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Получить список чатов
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: Создает новый чат с указанными участниками и настройками
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: Данные для создания чата
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/dto.ChatCreateInformationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: ID созданного чата
          schema:
            $ref: '#/definitions/dto.IdDTO'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Создать новый чат
      tags:
      - chats
...
  /chats/{chatId}:
    delete:
      consumes:
      - application/json
      description: Удаляет существующий чат. Только администратор чата может удалить
        чат.
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: ID чата
        format: uuid
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет прав для удаления чата
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Чат не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Удалить чат
      tags:
      - chats
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о конкретном чате, включая сообщения
        и участников
      parameters:
      - description: ID чата
        format: uuid
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Детальная информация о чате
          schema:
            $ref: '#/definitions/dto.ChatDetailedInformationDTO'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Не существует такого чата
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Получить информацию о чате
      tags:
      - chats
    patch:
      consumes:
      - application/json
      description: Позволяет администратору изменить название и описание чата
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: ID чата
        format: uuid
        in: path
        name: chatId
        required: true
        type: string
      - description: Поля для обновления чата
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/dto.ChatUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет прав для изменения чата
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Чат не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Обновить чат
      tags:
      - chats
  /chats/{chatId}/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Загружает новый аватар для чата (только админ)
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: ID чата
        format: uuid
        in: path
        name: chatId
        required: true
        type: string
      - description: Файл аватара
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: URL загруженного аватара
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "403":
          description: Нет прав для изменения чата
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Чат не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Загрузить аватарку чата
      tags:
      - chats
  /chats/{chatId}/members:
    patch:
      consumes:
      - application/json
      description: Добавляет указанных пользователей в существующий чат
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: ID чата
        format: uuid
        in: path
        name: chatId
        required: true
        type: string
      - description: Список пользователей для добавления
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/dto.AddUsersToChatDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Добавить пользователей в чат
      tags:
      - chats
  /chats/avatars/query:
    post:
      consumes:
      - application/json
      description: Возвращает аватарки для списка чатов по их ID
      parameters:
      - description: Список ID чатов
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetAvatarsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Аватарки чатов
          schema:
            $ref: '#/definitions/dto.GetAvatarsResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Получить аватарки чатов
      tags:
      - chats
  /chats/dialog/{otherUserId}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о личном чате между авторизованным пользователем
        и указанным пользователем. Если чата нет — возвращается ошибку.
      parameters:
      - description: ID другого пользователя
        format: uuid
        in: path
        name: otherUserId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о личном чате
          schema:
            $ref: '#/definitions/dto.IdDTO'
        "400":
          description: Некорректный ID пользователя
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Получить личный диалог с пользователем
      tags:
      - chats
  /chats/search:
    get:
      consumes:
      - application/json
      description: Позволяет искать чаты по части или полному имени
      parameters:
      - description: Имя или часть имени чата для поиска
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список найденных чатов
          schema:
            items:
              $ref: '#/definitions/dto.ChatViewInformationDTO'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/dto.ErrorDTO'
      security:
      - ApiKeyAuth: []
      summary: Поиск чатов по имени
      tags:
      - chats
...
swagger: "2.0"