# internal/transport/dto/chats/chats.go
package dto

import (
	dto "github.com/go-park-mail-ru/2025_2_Undefined/internal/transport/dto/message"
	"github.com/google/uuid"
)

type ChatViewInformationDTO struct {
	ID          uuid.UUID      `json:"id" swaggertype:"string" format:"uuid"`
	Name        string         `json:"name"`
	LastMessage dto.MessageDTO `json:"last_message" swaggertype:"object"`
	Type        string         `json:"type"`
}

type ChatDetailedInformationDTO struct {
	ID          uuid.UUID         `json:"id" swaggertype:"string" format:"uuid"`
	Name        string            `json:"name"`
	Description string            `json:"description"`
	IsAdmin     bool              `json:"is_admin"`
	CanChat     bool              `json:"can_chat"`
	IsMember    bool              `json:"is_member"`
	IsPrivate   bool              `json:"is_private"`
	Type        string            `json:"type"`
	Messages    []dto.MessageDTO  `json:"messages"`
	Members     []UserInfoChatDTO `json:"members"`
}

type ChatCreateInformationDTO struct {
	Name    string             `json:"name"`
	Type    string             `json:"type"` // Тип чата - канал, диалог или группа
	Members []AddChatMemberDTO `json:"members"`
}

type UserInfoChatDTO struct {
	UserId   uuid.UUID `json:"user_id" swaggertype:"string" format:"uuid"`
	UserName string    `json:"user_name"`
	Role     string    `json:"role"` // Роль пользователя в чате - админ(писать и добавлять участников), участник(писать), зритель (только просмотр)
}

type AddChatMemberDTO struct {
	UserId uuid.UUID `json:"user_id" swaggertype:"string" format:"uuid"`
	Role   string    `json:"role"` // Роль пользователя в чате - админ(писать и добавлять участников), участник(писать), зритель (только просмотр)
}

type AddUsersToChatDTO struct {
	Users []AddChatMemberDTO `json:"members"`
}

type ChatUpdateDTO struct {
	Name        string `json:"name"`
	Description string `json:"description"`
}